Rolled back optimization/trimming due to BUG
A recently executed test showed that when approaching the convergence
point of a problem, the main energy stooped being monotonously
decreasing and instead it increased by small quantities.
The possible explanation that the measure_trimming routine was modified
to discard current_measure.curves in favor of stationary_curves
when two are "similar enough to consider them to be equal".